+++ CHANGELOG.md
@@ -637,6 +637,7 @@
 * Removed the AttributeError warning from console when a player inputs a command that does not exist
 * Added the command name that was used whenever an "Invalid command" error message is triggered
 * Made /sneak and /reveal (/unsneak) with no arguments affect the player using it rather than raising an error
+* Added `source` optional parameter to music list files, which would indicate the source of the currently played music via /currentmusic
 * Fixed scream_range in area list yaml files not supporting the keyword <ALL> to indicate all areas should be able to receive a scream coming from a particular area
 * Fixed scream_range in area list yaml files not checking if the areas a scream can reach to from a particular area exist
 * Fixed /scream, /whisper and /party_whisper raising errors if a message was sent to a deafened player with a bypass message starter. They now sent messages but filtered